requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
redis>=5.0.0
SQLAlchemy>=2.0.0
asyncmy>=0.2.9
PyJWT>=2.8.0
pydantic>=2.0.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
cryptography>=41.0.0
bcrypt==3.2.0
